Refactor: 상수파일 config 폴더로 위치 변경과 그에 따른 import 주소 수정
